requirements.txt
pandas==2.3.3
plotly==6.6.0
requests==2.32.5
streamlit==1.55.0
yfinance==1.2.0
openai>=1.0.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1
